requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy>=2.0.25
Werkzeug
python-dotenv
psycopg2-binary
bcrypt
gunicorn
